Clean up the plugin_template folder, and add a new simple example.
