--- plugin_template/src/examples/HelloReact.tsx
@@ -1,39 +1,0 @@
-import React from "react";
-import "../App.css";
-import StreamlitPlugin from "../PluginWrapper";
-import { StComponentProps } from "../StComponentAPI";
-import logo from "../logo.svg";
-
-/**
- * Hello World example plugin, using React.
- */
-class HelloReact extends React.PureComponent<StComponentProps> {
-  public render = (): React.ReactNode => {
-    return (
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <p>Hello, {`${this.props.args["name"]}`}!</p>
-          <a
-            className="App-link"
-            href="https://reactjs.org"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Learn React
-          </a>
-        </header>
-      </div>
-    );
-  };
-
-  public componentDidUpdate = (): void => {
-    this.props.updateFrameHeight();
-  };
-
-  public componentDidMount = (): void => {
-    this.props.updateFrameHeight();
-  };
-}
-
-export default StreamlitPlugin(HelloReact);
